Broke Time into it's own folder. Added some more styling updates
